channelmodel: cast QObject::sender to QHtspEvent as that's his type
Fix a bug where the row was always -1. This was because the QObject::sender was
successfully casted to a QHtspChannel, while in reality it was a QHtspEvent.
This caused the indexOf on the channellist to always return -1. The fix is
obviously to cast the sender to the propper QHtspEvent and to get the channel
from the event.